Add files via upload
A dynamic student form built using HTML, CSS, and JavaScript, allowing users to input and manage student data. The form captures details such as name, favorite color, gender, and country, and stores the information in local storage. The app features functionality to display the stored data in a table format, update entries, and refill the form for editing. Designed to enhance user interaction and provide a seamless data management experience.